(HtmlElement/GymPriceInfo)Type + BeActiveParser
Created the parser class. BeActiveParser is configured to correctly parse the
BeActive Gym prices.

The HtmlElementType stores the data about the desired html elements:
xpath, name of the purpose and the value of the parsed
html element.

GymPriceInfoType is supposed to contain the calculated and fetched data
of the described gym.

Temporarily the BeActiveParser contains the function to map the specific
params to the GymPriceInfoType.